Move populate() +  make population methods private
populate() was moved out of init_db(), but placed in init_db_command(). This is so that we can populate the database when we initialize a database instance on a new machine, but don't populate the test instance of the db with the full csv of data. Also the populate() and its helpers were marked as private since they shouldn't be used outtside of this module.