Fix query tests (#126)
* Experiment: Remove app.db lines from conftest

They don't mathc up with the flaskr docs, and Chase added them before I fixed the html pages... hmm let's see if it works

* add footnote & status to git_tips.md

* remove line from ci.yaml

* try removing execute fake data script line

dude i swear

* test if test db works

* Add test_db_exist

I hate this and nothing makes sense

* Coverage testing (#116)

Everything looks good. All current tests pass, we need to improve our coverage but that's something we know because of this branch.

* Move populate() +  make population methods private

populate() was moved out of init_db(), but placed in init_db_command(). This is so that we can populate the database when we initialize a database instance on a new machine, but don't populate the test instance of the db with the full csv of data. Also the populate() and its helpers were marked as private since they shouldn't be used outtside of this module.

* Add TODO list for Eric in conftest.py

* Add test_test_db()

Eric, I think you can use something like this to check if the test instance of the db is populated?

* Copied database schema from working schema and added INSERT statements to fill it with testable data

* renamed imported _populate_db line to fix error

* Fixed type

* Changed city inserts and added comment

* Added test for the populate_db function

* Added test for user table in test_db

* Removed commented out lines when initializing test database

* Removed TODO comment

* Update test_queries.py with new test DB expected data

* Change data.sql - replace booleans with yes/no

---------

Co-authored-by: Eric Harrison <hericeh72@gmail.com>