Updated Testing
Updated auth_actions to accurately reflect what is in the test db. My branch added a few users to more adequately test a few functions that I wrote, so the previous default user id was clashing with a user ID that was already in the db. fixed it so that it actually creates a new, unique user id